Font Rendering (#73)
* Add custom pixel font texture, with the regions manually worked out in `ui.cpp`. Bit more time consuming that using freetype, but have completely control over it.
* Renamed `texture` to `texture_dyn` and add `texture_png`. The class was getting overly complicated with different config settings, so I just made them separate. `texture_dyn` was chosen at the name as the texture is dynamically generated in the code.
* Signed its are used for dimenions everywhere now, removes the need for casting in a lot of places

+++ src/editor.m.cpp
@@ -61,7 +61,7 @@ auto main() -> int
     auto editor          = sand::editor{};
     auto input           = sand::input{};
     auto level           = sand::new_level(4, 4);
-    auto world_renderer  = sand::renderer{static_cast<u32>(level->pixels.width_in_pixels()), static_cast<u32>(level->pixels.height_in_pixels())};
+    auto world_renderer  = sand::renderer{level->pixels.width_in_pixels(), level->pixels.height_in_pixels()};
     auto accumulator     = 0.0;
     auto timer           = sand::timer{};
     auto shape_renderer  = sand::shape_renderer{};
